Move 3 fixtures to ixmp.testing
